local-rag-system/requirements.txt
# Local RAG System Dependencies
# =============================

# LLM Backend - llama-cpp-python for running GGUF quantized models
llama-cpp-python>=0.2.20

# Embeddings and Vector Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Text Processing
langchain>=0.1.0
langchain-community>=0.0.10
tiktoken>=0.5.1

# CLI and UX
rich>=13.7.0
click>=8.1.7
prompt_toolkit>=3.0.43

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.1
pyyaml>=6.0.1

# Testing and Evaluation
pytest>=7.4.3
rouge-score>=0.1.2
nltk>=3.8.1

# Dataset Loading (Hugging Face)
datasets>=2.14.0
huggingface-hub>=0.17.0

